+++ lib/gql/honeycomb.middleware.d.ts
@@ -2,8 +2,6 @@ import { NextFn, ResolverData, MiddlewareInterface } from 'type-graphql';
 export declare class HoneycombMiddleware implements MiddlewareInterface {
     use({ root, args, context, info }: ResolverData, next: NextFn): any;
 }
-/**
- * check root for span
- * create span with parentid from root span
- * set span on current root
- */
+export declare class RumbleshipTraceMiddleware implements MiddlewareInterface {
+    use({ root, args, context, info }: ResolverData, next: NextFn): any;
+}