+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from './rfi-beeline';
 export * from './gql/honeycomb.middleware';
 export * from './with-span.decorator';
+export * from './rumbleship-beeline';
+export * from './honeycomb.interfaces';