+++ src/with-span.decorator.ts
@@ -1,5 +1,10 @@
 import 'reflect-metadata';
+import { RumbleshipBeeline } from './rumbleship-beeline';
 export function WithSpan(context: object = {}): MethodDecorator {
+  // tslint:disable-next-line: no-console
+  console.warn(
+    '@rumbleship/o11y.WithSpan is now deprecated. Update to @rumbleship/o11y.WithAsyncSpan'
+  );
   return (target: any, propertyName: string | symbol, descriptor: PropertyDescriptor) => {
     const originalMethod = descriptor.value;
     // NOTE: because the ctx is defined in the framework, this should probably be defined in the framework too
@@ -26,4 +31,52 @@ export function WithSpan(context: object = {}): MethodDecorator {
       });
     };
   };
+}
+
+/**
+ *
+ * @param span_metadata?: {}  metadata to add to the Honeycomb trace context.
+ * @default context {
+ *                    name: `${class.constructor.name}.${wrappedMethod.name}`
+ *                    class: `${class.constructor.name}`,
+ *                    method: `${wrappedMethod.name}`
+ *                  }
+ */
+export function AddToTrace(span_metadata: object = {}): MethodDecorator {
+  return (target: any, propertyName: string | symbol, descriptor: PropertyDescriptor) => {
+    if (!(span_metadata as any).name) {
+      Reflect.set(span_metadata, 'name', `${target.constructor.name}.${propertyName.toString()}`);
+    }
+    Reflect.set(span_metadata, 'class', target.constructor.name);
+    Reflect.set(span_metadata, 'method', propertyName.toString());
+    const originalMethod = descriptor.value;
+    descriptor.value = function(...args: any[]) {
+      const { beeline } = (this as any).ctx || findContextWithBeelineFrom(args) || {};
+      if (beeline) {
+        const spanContext = {
+          'origin.type': 'decorator',
+          ...span_metadata
+        };
+
+        const wrapped = () => originalMethod.apply(this, args);
+
+        return beeline.bindFunctionToTrace(() => {
+          return beeline.withAsyncSpan(spanContext, wrapped);
+        });
+      } else {
+        throw new Error(
+          `Cannot find an RumbleshipContext. Two solutions: 
+            1. Pass it as an argument,
+            2. Decorate a method of a RumbleshipService that has \`.ctx\` property already assigned`
+        );
+      }
+    };
+  };
+}
+
+// I don't want to introduce a cyclic dependency to `@rumbleship/gql` to be able
+// do this check with a plain `instanceof RumbleshipContext` -- so we check for a feature we know
+// and very much _need_ to be present: the beeline.
+function findContextWithBeelineFrom(args: any[]): { beeline: RumbleshipBeeline } | undefined {
+  return args.find(arg => !!(arg.beeline instanceof RumbleshipBeeline));
 }