+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * Typesafe `RumbleshipBeeline` that acts as its own factory, enabling tracking of fullly async service contexts
+  * `@AddToTrace` decorator to bind function calls into the trace
+### Removed
+### Changed
+  * Updated tslint.json, tsconfig.json, plugins, etc to support nullish coalescing.
+### Fixed
+### Deprecated
+  * `src/rfi-beeline.ts`: `RFIBeeline` and `HoneycombBeelineFactory` 
+### Security
+
+
 ## [0.0.7] -- 2020-01-24
 
 ### Fixed